refactor: Created new producer to encapsulate creating of items
