+++ matchingService/src/QueueService/AmqpService.ts
@@ -1,12 +1,21 @@
 import { Channel } from "amqplib";
 import QueueMessage from "./QueueMessage";
 import MatchRequest from "./MatchRequest";
-import { MessageHeader, ConsumerMessageHeaderReq } from "./MessageHeaders";
 import { ConnectionManager, IConnectionManager } from "./ConnectionManager";
 import ChannelNotFoundError from "../errors/ChannelNotFoundError";
+import Consumer from "./Consumer";
+import Producer from "./Producer";
 
 const TOPIC_LIST: string[] = ["algorithm", "graph", "dp"];
 const DIFFICULTY_LEVELS: string[] = ["easy", "medium", "hard"];
+
+const testMessages: MatchRequest[] = [
+    new MatchRequest("john", "algorithm", "hard"),
+    new MatchRequest("amy", "algorithm", "medium"),
+    new MatchRequest("johhny", "graph", "easy"),
+    new MatchRequest("bob", "graph", "easy"),
+    new MatchRequest("the builder", "dp", "medium"),
+];
 
 class AmqpService {
     private exchange: string;
@@ -34,53 +43,28 @@ class AmqpService {
         return service;
     }
 
-    public async sendJsonMessage(msg: MatchRequest): Promise<void> {
+    public async startProducers(): Promise<void> {
         var channel: Channel = this.connectionManager.getChannel();
         if (channel instanceof ChannelNotFoundError) {
             console.error(channel.message);
             return;
         }
-        const messageHeaders: MessageHeader = {
-            topic: msg.getTopic(),
-            difficulty: msg.getDifficulty()
+        var producer: Producer = new Producer();
+        for (const testMessage of testMessages) {
+            await producer.sendJsonMessage(testMessage, channel, this.exchange);
         }
-        channel.publish(this.exchange, "", Buffer.from(JSON.stringify(msg)), {
-            headers: messageHeaders
-        });
-        console.log("Successfully sent message!");
-    }
-
-    public async receiveMessages(topic: string, difficulty: string, allback: (msg: QueueMessage | null) => void): Promise<void> {
-        var channel: Channel = this.connectionManager.getChannel();
-        if (channel instanceof ChannelNotFoundError) {
-            console.error(channel.message);
-            return;
-        }
-        const queue = await channel.assertQueue("", { exclusive: true })
-        const queueName = queue?.queue;
-        if (!queueName) {
-            return;
-        }
-
-        var headers: ConsumerMessageHeaderReq = {
-            "x-match": "all",
-            "topic": topic,
-            "difficulty": difficulty
-        }
-
-        await channel.bindQueue(queueName, this.exchange, "", headers);
-        console.log("Consumer ready");
-        await channel.consume(queueName, (message) => {
-            if (message) {
-                console.log(`Received message: ${message.content.toString()}`);
-            }
-        }, { noAck: true });
     }
 
     public async startConsumers(): Promise<void> {
+        var channel: Channel = this.connectionManager.getChannel();
+        if (channel instanceof ChannelNotFoundError) {
+            console.error(channel.message);
+            return;
+        }
+        var consumer: Consumer = new Consumer();
         for (const topic of TOPIC_LIST) {
             for (const difficulty of DIFFICULTY_LEVELS) {
-                await this.receiveMessages(topic, difficulty, (msg: QueueMessage | null) => {
+                await consumer.receiveMessages(topic, difficulty, this.exchange, channel, (msg: QueueMessage | null) => {
                     let content = msg?.content.toString();
                     if (content) {
                         try {
